# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CVE Atlas API
  version: 0.1.0
  summary: Foundation API for CVE Atlas
  description: >-
    Milestone 0 foundation API. It contains platform health, OIDC session handling,
    RBAC-protected audit access, and no active scanning capabilities.
servers:
  - url: http://localhost:8080
    description: Local development
security: []
tags:
  - name: Health
  - name: Authentication
  - name: Audit
paths:
  /health/live:
    get:
      tags: [Health]
      operationId: getLiveness
      summary: Process liveness
      responses:
        "200":
          description: Process is alive.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LivenessResponse"
  /health/ready:
    get:
      tags: [Health]
      operationId: getReadiness
      summary: Dependency readiness
      responses:
        "200":
          description: Required dependencies are available.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
        "503":
          description: One or more required dependencies are unavailable.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
  /api/v1/auth/login:
    get:
      tags: [Authentication]
      operationId: startOIDCLogin
      summary: Start OIDC Authorization Code flow with PKCE
      parameters:
        - name: return_to
          in: query
          required: false
          description: Same-origin web URL or relative path used after successful login.
          schema:
            type: string
            maxLength: 2048
      responses:
        "302":
          description: Redirect to the configured OIDC provider.
          headers:
            Location:
              schema:
                type: string
                format: uri
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/callback:
    get:
      tags: [Authentication]
      operationId: completeOIDCLogin
      summary: Complete OIDC Authorization Code flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Authentication succeeded and a secure application session was created.
          headers:
            Location:
              schema:
                type: string
                format: uri
            Set-Cookie:
              schema:
                type: string
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentPrincipal
      summary: Return the authenticated principal and effective authorization
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Authenticated principal.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Principal"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logoutCurrentSession
      summary: Revoke the current application session
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Session revoked.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [logout_url]
                    properties:
                      logout_url:
                        type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/audit-events:
    get:
      tags: [Audit]
      operationId: listAuditEvents
      summary: List append-only tenant audit events
      description: Requires the audit.read permission.
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Audit event page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
                  pagination:
                    $ref: "#/components/schemas/OffsetPagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cve_atlas_session
  headers:
    RequestID:
      description: Request correlation identifier.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._:-]{1,128}$"
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Unauthorized:
      description: Authentication is required or the session is invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The current principal is not allowed to access the resource.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An internal error occurred. Stack traces are never returned.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    LivenessResponse:
      type: object
      required: [status, time]
      properties:
        status:
          type: string
          const: alive
        time:
          type: string
          format: date-time
    DependencyStatus:
      type: object
      required: [status, latency_ms]
      properties:
        status:
          type: string
          enum: [ok, unavailable]
        latency_ms:
          type: integer
          minimum: 0
    ReadinessResponse:
      type: object
      required: [status, checked_at, dependencies]
      properties:
        status:
          type: string
          enum: [ready, not_ready]
        checked_at:
          type: string
          format: date-time
        dependencies:
          type: object
          required: [postgres, redis, neo4j, object_storage]
          additionalProperties:
            $ref: "#/components/schemas/DependencyStatus"
    Principal:
      type: object
      required: [user_id, tenant_id, roles, permissions]
      properties:
        user_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        preferred_username:
          type: string
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
        permissions:
          type: array
          uniqueItems: true
          items:
            type: string
    AuditEvent:
      type: object
      required: [id, occurred_at, tenant_id, action, result, metadata]
      properties:
        id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date-time
        actor_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        result:
          type: string
          enum: [success, failure, denied]
        request_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
    OffsetPagination:
      type: object
      required: [limit, offset, next_offset]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        next_offset:
          type: integer
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"